requirements.txt
# MCP SDK
mcp>=1.17.0

# Async HTTP client for scanning
aiohttp>=3.9.0
httpx>=0.25.0

# Network scanning
python-nmap>=0.7.1

# Security analysis
cryptography>=41.0.0
pydantic>=2.5.0

# CLI interface
click>=8.1.0
rich>=13.7.0
tabulate>=0.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Reporting
jinja2>=3.1.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.12.0
ruff>=0.1.9
mypy>=1.7.0
